// Qanat.Web/src/app/shared/generated/model/usage-location-update-cover-crop-dto.ts
/**
 * Qanat.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class UsageLocationUpdateCoverCropDto { 
    UsageLocationTypeID: number;
    CoverCropNote?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface UsageLocationUpdateCoverCropDtoForm { 
    UsageLocationTypeID: FormControl<number>;
    CoverCropNote?: FormControl<string>;
}

export class UsageLocationUpdateCoverCropDtoFormControls { 
    public static UsageLocationTypeID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: true,
            validators: 
            [
                Validators.required,
            ],
        }
    );
    public static CoverCropNote = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}